+++ gear/form_qc_coordinator/src/python/form_qc_coordinator_app/coordinator.py
@@ -11,6 +11,10 @@ from configs.ingest_configs import (
     SupplementModuleConfigs,
 )
 from dates.form_dates import DEFAULT_DATE_TIME_FORMAT
+from error_logging.error_logger import (
+    reset_error_log_metadata_for_gears,
+    update_error_log_and_qc_metadata,
+)
 from flywheel.models.acquisition import Acquisition
 from flywheel.models.file_entry import FileEntry
 from flywheel.rest import ApiException
@@ -35,17 +39,13 @@ from nacc_common.error_models import (
     VisitKeys,
 )
 from nacc_common.field_names import FieldNames
-from outputs.error_logger import (
-    get_file_qc_info,
-    reset_error_log_metadata_for_gears,
-    update_error_log_and_qc_metadata,
-)
 from outputs.error_writer import ListErrorWriter
 from outputs.errors import (
     preprocessing_error,
     previous_visit_failed_error,
     system_error,
 )
+from pydantic import ValidationError
 
 from form_qc_coordinator_app.visits import VisitsLookupHelper
 
@@ -164,8 +164,12 @@ class QCCoordinator:
             data=error_writer.errors().model_dump(by_alias=True),
         )
         visit_file = visit_file.reload()
-        qc_info = get_file_qc_info(visit_file)
-        if not qc_info:
+        try:
+            qc_info = FileQCModel.create(visit_file)
+        except ValidationError as error:
+            log.warning(
+                "Error loading QC metadata for file %s: %s", visit_file.name, error
+            )
             qc_info = FileQCModel(qc={})
 
         # add qc-coordinator gear info to visit file metadata
@@ -226,8 +230,12 @@ class QCCoordinator:
             error_writer.write(error_obj)
 
         visit_file = visit_file.reload()
-        qc_info = get_file_qc_info(visit_file)
-        if not qc_info:
+        try:
+            qc_info = FileQCModel.create(visit_file)
+        except ValidationError as error:
+            log.warning(
+                "Error loading QC metadata for file %s: %s", visit_file.name, error
+            )
             qc_info = FileQCModel(qc={})
 
         # rest form-qc-checker errors if any