+++ gear/identifier_lookup/test/python/test_identifier_app.py
@@ -10,7 +10,7 @@ from identifiers.model import IdentifierObject
 from nacc_common.error_models import FileError
 from outputs.error_writer import ListErrorWriter
 from test_mocks.mock_configs import uds_ingest_configs
-from test_mocks.mock_flywheel import MockProject
+from test_mocks.mock_flywheel import MockProjectAdaptor
 
 
 @pytest.fixture(scope="function")
@@ -134,7 +134,7 @@ class TestIdentifierLookup:
                 error_writer=error_writer,
                 gear_name="dummy",
                 misc_errors=misc_errors,
-                project=MockProject(label="dummy-project"),
+                project=MockProjectAdaptor(label="dummy-project"),
             ),
             error_writer=error_writer,
         )
@@ -160,7 +160,7 @@ class TestIdentifierLookup:
                 error_writer=error_writer,
                 gear_name="dummy",
                 misc_errors=misc_errors,
-                project=MockProject(label="dummy-project"),
+                project=MockProjectAdaptor(label="dummy-project"),
             ),
             error_writer=error_writer,
         )
@@ -186,7 +186,7 @@ class TestIdentifierLookup:
                 error_writer=error_writer,
                 gear_name="dummy",
                 misc_errors=misc_errors,
-                project=MockProject(label="dummy-project"),
+                project=MockProjectAdaptor(label="dummy-project"),
             ),
             error_writer=error_writer,
         )
@@ -212,7 +212,7 @@ class TestIdentifierLookup:
                 error_writer=error_writer,
                 gear_name="dummy",
                 misc_errors=misc_errors,
-                project=MockProject(label="dummy-project"),
+                project=MockProjectAdaptor(label="dummy-project"),
             ),
             error_writer=error_writer,
         )
@@ -246,7 +246,7 @@ class TestIdentifierLookup:
                 error_writer=error_writer,
                 gear_name="dummy",
                 misc_errors=misc_errors,
-                project=MockProject(label="dummy-project"),
+                project=MockProjectAdaptor(label="dummy-project"),
             ),
             error_writer=error_writer,
         )