+++ docs/form_qc_coordinator/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this gear are documented in this file.
 
+## Unreleased
+* Integrates event logging functionality for enhanced QC status tracking and error reporting
+
 ## 1.4.1
 * Fixes a bug in updating the visit error log